Python/Codechef_Problem_Scrapper/requirements.txt
Pillow==7.2.0
selenium==3.141.0
urllib3==1.25.10
